Remove `const` from value type references
There are several places where we were passing around pointers to
`const struct vrt_value_type` instances.  However, we can't really
require the value type implementations to not be mutable; it might be
necessary for a value type to update one of its fields whenever an
value instance is created or freed.