Configurable starting queue index
Each producer and consumer uses an integer index to identify the next
slot in the ring buffer to write into or read from.  Before, we used a
very large initial index, to make sure that the index overflows (and
wraps around to 0) at some point during the test.  This makes the
debugging messages hard to read during normal operation, however, since
the numbers all look pretty similar.  Now, the default is to start each
queue index at 0 (to make them easy to read), and provide a function
(`vrt_testing_mode`) to use the high starting value in test cases.