Debug messages using clogger
All of the existing debug messages (which are normally ifdef-ed out)
have been replaced with clogger debug messages.  Most of them are
trace-level messages, but we have debug-level messages that can be used
to see when producers and consumers have to yield.  This should
(hopefully) be useful in detecting deadlocks.